refactor: 동적 메타데이터 적용
