+++ public/data/products/5.json
@@ -1,5 +1,6 @@
 {
   "name": "사용자 맞춤 클래식 공연 큐레이션 서비스, Claco",
+  "shortName": "Claco",
   "category": "큐레이팅",
   "introduction": "최선의 클래식 공연 선택을 위한 길잡이",
   "description": "Claco는 클래식 공연을 즐기고 싶은 모든 사람을 위한 개인 맞춤형 큐레이션 서비스입니다.\n클래식 공연을 처음 접하는 사람부터 애호가까지, 자신만의 클래식 취향을 발견하고 공연을 선택하는 과정을 쉽고 즐겁게 만들어줍니다. Claco는 사용자 데이터를 바탕으로 AI가 취향에 맞는 공연을 추천하고, 공연 선택부터 감상 후 기록과 공유까지의 여정을 지원합니다.\n사용자는 공연 감상을 기록해 티켓북에 소장하고, 추억을 간직하고 공유할 수 있습니다.\nClaco와 함께라면, 클래식은 더 이상 어렵지 않습니다.\n최선의 클래식 공연 선택을 위한 길잡이 Claco와 당신만의 클래식 여정을 시작하세요!",
