+++ public/data/products/6.json
@@ -1,5 +1,6 @@
 {
   "name": "모일 때 맵핀, MOPING",
+  "shortName": "MOPING",
   "category": "커뮤니티",
   "introduction": "함께 방문하는 사람들의 북마크를 모아, 우리가 만드는 공간 공유 지도",
   "description": "MOPING 은 함께 방문하는 사람들의 북마크를 모아주는 공간 공유 지도 서비스입니다. 내가 신뢰하는 친구의 북마크를 통해 믿을만한 '느좋 공간' 의 정보를 알아볼 수 있습니다. 또한, 모임 별로 공간 정보가 축적되어 매번 새롭게 찾아볼 필요도 없죠.\n어디갈지 고민하며 일일이 가게 정보를 공유하던 과정, 번거롭지 않으셨나요? MOPING 으로 네가 고른 곳•내가 고른 곳 모두 모아 한 번에 확인하고 편리하게 결정하세요!",
