+++ public/data/products/8.json
@@ -1,5 +1,6 @@
 {
   "name": "일단 | 일일계획을 단순하게",
+  "shortName": "일단",
   "category": "생산성",
   "introduction": "일단 ‘할 일’ 쌓고 ‘오늘’에 집중할 수 있도록 돕는 앱",
   "description": "일정 관리, 잘 하고 계신가요?\n계획대로 하루를 보내지 못하면 부정적인 감정이 마음속을 사로잡곤 하죠. 그래서 생각했어요. ‘계획 수립과 실패, 그리고 실망… 반복되는 사이클을 끊어낼 수 없을까?‘ ‘일단’은 수많은 일들 중 ‘오늘’ 할 일에 집중할 수 있도록 돕는 앱이에요.\n고민은 뒤로하고, 일단 해야 할 일을 무작정 쌓아 보세요. 그리고 오늘 해야 할 일만 간단하게 스와이프하면 손쉽게 투두리스트가 완성되죠.\n오늘 해야 할 일을 못하셨다고요? 괜찮아요. “일단”과 함께 다시 시작하면 되니까요!",
