+++ src/app/archive/layout.tsx
@@ -1,6 +1,14 @@
 import { Suspense } from 'react'
 
+import { Metadata } from 'next'
+
 import ArchiveLayoutContent from '@src/containers/archive/LayoutContent'
+
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: 'ARCHIVE',
+  }
+}
 
 export default function ArchiveLayout({
   children,
