+++ src/app/layout.tsx
@@ -1,12 +1,15 @@
+import { Metadata } from 'next'
+
 import { Header } from '@src/components/Header'
-import { metadata } from '@src/constants/metadata'
+import { baseMetadata } from '@src/constants/metadata'
 import { AboutSectionRefProvider } from '@src/contexts/AboutSectionRefContext'
 import { GuestBookProvider } from '@src/contexts/GuestBookContext'
 import { pretendard } from '@src/styles/font'
-
 import '@src/styles/globals.css'
 
-export { metadata }
+export async function generateMetadata(): Promise<Metadata> {
+  return baseMetadata
+}
 
 export default function RootLayout({
   children,
