+++ src/app/people/layout.tsx
@@ -1,6 +1,14 @@
 import { Suspense } from 'react'
 
+import { Metadata } from 'next'
+
 import PeopleLayoutContent from '@src/containers/people/LayoutContent'
+
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: 'PEOPLE',
+  }
+}
 
 export default function PeopleLayout({
   children,
